refactor: Move rental detail mappers to data layer (#121)
Moves `OwnerRentalDetailMapper.kt` and `RenterRentalDetailMapper.kt` from the presentation layer (`presentation.rentaldetail.owner/renter`) to the data layer (`data.rental.mapper`).
