+++ app/src/main/java/com/example/rentit/data/rental/mapper/RenterRentalDetailMapper.kt
@@ -1,4 +1,4 @@
-package com.example.rentit.presentation.rentaldetail.owner
+package com.example.rentit.data.rental.mapper
 
 import android.os.Build
 import androidx.annotation.RequiresApi
@@ -7,10 +7,10 @@ import com.example.rentit.common.util.daysFromToday
 import com.example.rentit.data.rental.dto.RentalDetailResponseDto
 import com.example.rentit.common.uimodel.RentalSummaryUiModel
 import com.example.rentit.presentation.rentaldetail.model.RentingStatus
-import com.example.rentit.presentation.rentaldetail.owner.stateui.OwnerRentalStatusUiModel
+import com.example.rentit.presentation.rentaldetail.renter.stateui.RenterRentalStatusUiModel
 
 @RequiresApi(Build.VERSION_CODES.O)
-fun RentalDetailResponseDto.toOwnerUiModel(): OwnerRentalStatusUiModel {
+fun RentalDetailResponseDto.toRenterUiModel(): RenterRentalStatusUiModel {
     val status = rental.status
 
     val rentalSummary = RentalSummaryUiModel(
@@ -28,36 +28,20 @@ fun RentalDetailResponseDto.toOwnerUiModel(): OwnerRentalStatusUiModel {
         RentalStatus.ACCEPTED,
         RentalStatus.REJECTED,
         RentalStatus.CANCELED -> {
-            OwnerRentalStatusUiModel.Request(
+            RenterRentalStatusUiModel.Request(
                 status = status,
-                isPending = status == RentalStatus.PENDING,
                 isAccepted = status == RentalStatus.ACCEPTED,
                 rentalSummary = rentalSummary,
                 basicRentalFee = basicRentalFee,
+                deposit = rental.depositAmount
             )
         }
 
         RentalStatus.PAID -> {
             val daysUntilRental = daysFromToday(rental.startDate)
-            OwnerRentalStatusUiModel.Paid(
+            RenterRentalStatusUiModel.Paid(
                 status = status,
                 daysUntilRental = daysUntilRental,
-                rentalSummary = rentalSummary,
-                basicRentalFee = basicRentalFee,
-                isSendingPhotoRegistered = rental.deliveryStatus.isPhotoRegistered,
-                isSendingTrackingNumRegistered = rental.deliveryStatus.isTrackingNumberRegistered,
-                rentalTrackingNumber = rental.rentalTrackingNumber,
-            )
-        }
-
-        RentalStatus.RENTING -> {
-            val daysFromReturnDate = daysFromToday(rental.endDate)
-            val rentingStatus = RentingStatus.fromDaysFromReturnDate(daysFromReturnDate)
-
-            OwnerRentalStatusUiModel.Renting(
-                status = rentingStatus,
-                isOverdue = rentingStatus == RentingStatus.RENTING_OVERDUE,
-                daysFromReturnDate = daysFromReturnDate,
                 rentalSummary = rentalSummary,
                 deposit = rental.depositAmount,
                 basicRentalFee = basicRentalFee,
@@ -65,7 +49,26 @@ fun RentalDetailResponseDto.toOwnerUiModel(): OwnerRentalStatusUiModel {
             )
         }
 
-        RentalStatus.RETURNED -> OwnerRentalStatusUiModel.Returned(
+        RentalStatus.RENTING -> {
+            val daysFromReturnDate = daysFromToday(rental.endDate)
+            val rentingStatus = RentingStatus.fromDaysFromReturnDate(daysFromReturnDate)
+            val isReturnAvailable = daysFromReturnDate <= -1
+
+            RenterRentalStatusUiModel.Renting(
+                status = rentingStatus,
+                isOverdue = rentingStatus == RentingStatus.RENTING_OVERDUE,
+                isReturnAvailable = isReturnAvailable,
+                daysFromReturnDate = daysFromReturnDate,
+                rentalSummary = rentalSummary,
+                deposit = rental.depositAmount,
+                basicRentalFee = basicRentalFee,
+                isReturnPhotoRegistered = rental.returnStatus.isPhotoRegistered,
+                isReturnTrackingNumRegistered = rental.returnStatus.isTrackingNumberRegistered,
+                rentalTrackingNumber = rental.rentalTrackingNumber
+            )
+        }
+
+        RentalStatus.RETURNED -> RenterRentalStatusUiModel.Returned(
             status = status,
             rentalSummary = rentalSummary,
             deposit = rental.depositAmount,
